Use 3 rolls to score in the final frame
This is the last frame, so using all remaining rolls made sense
initially, but there will never be more than 3 rolls in a final frame.
Even if the final frame is a strike, the player only gets 2 more "bonus"
rolls. If these are both 10's, there is no additional bonus.

This is just to make this part of the logic more similar to the others.
Instead of having a dynamic # of rolls based on what's remaining, just
use 3 because that's the max. If there are fewer than 3 rolls in the
final frame, `drop` and `take` will just grab as many rolls as they can.